feat: add template smoke-test ci
