+++ {{ cookiecutter.project_slug }}/Dockerfile-python
@@ -6,8 +6,8 @@ ENV UV_PROJECT_ENVIRONMENT=/opt/venv
 ENV VIRTUAL_ENV=/opt/venv
 ENV PATH="$VIRTUAL_ENV/bin:$PATH"
 
-COPY pyproject.toml uv.lock ./
+COPY pyproject.toml ./
 RUN pip install --upgrade pip && pip install --no-cache-dir uv
-RUN uv sync --frozen --no-dev --no-install-project
+RUN uv sync --no-dev --no-install-project
 
 COPY . .
