+++ {{ cookiecutter.project_slug }}/apps/core/stripe_webhooks.py
