+++ {{ cookiecutter.project_slug }}/apps/core/tests/conftest.py
@@ -1,33 +1,6 @@
 import pytest
-from django.conf import settings
 
 from apps.core import tasks as core_tasks
-
-
-def pytest_configure(config):
-    settings.STORAGES["staticfiles"]["BACKEND"] = (
-        "django.contrib.staticfiles.storage.StaticFilesStorage"
-    )
-
-
-@pytest.fixture
-def user(django_user_model):
-    return django_user_model.objects.create_user(
-        username="testuser",
-        email="testuser@example.com",
-        password="password123",
-    )
-
-
-@pytest.fixture
-def auth_client(client, user):
-    client.force_login(user)
-    return client
-
-
-@pytest.fixture
-def profile(user):
-    return user.profile
 
 
 @pytest.fixture
