+++ {{ cookiecutter.project_slug }}/apps/core/tests/test_signals.py
