+++ {{ cookiecutter.project_slug }}/apps/core/tests/test_stripe_webhooks.py
