+++ {{ cookiecutter.project_slug }}/conftest.py
@@ -1,0 +1,28 @@
+import pytest
+from django.conf import settings
+
+
+def pytest_configure(config):
+    settings.STORAGES["staticfiles"]["BACKEND"] = (
+        "django.contrib.staticfiles.storage.StaticFilesStorage"
+    )
+
+
+@pytest.fixture
+def user(django_user_model):
+    return django_user_model.objects.create_user(
+        username="testuser",
+        email="testuser@example.com",
+        password="password123",
+    )
+
+
+@pytest.fixture
+def auth_client(client, user):
+    client.force_login(user)
+    return client
+
+
+@pytest.fixture
+def profile(user):
+    return user.profile
